+++ src/extension/extension.ts
@@ -59,7 +59,7 @@ export function activate(context: vs.ExtensionContext, isRestart: boolean = fals
 
 	// Use the console to output diagnostic information (console.log) and errors (console.error)
 	// This line of code will only be executed once when your extension is activated
-	console.log('Congratulations, your extension "Dart Polisher" is now active!');
+	console.log('"Dart Polisher" extension is now active!');
 
 	// Set up log files. (redirect category logs to these files)
 	setupLog(config.polisherLogFile, [LogCategory.Formatter]);