+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ## [Unreleased]
 ### New
 - New setting for logging verbose comunication between the extension and the local dart_polisher formatter server.
-- Web Extension! working (using dart2js, formatter server is not used), functionality is almost the same, (~3x slower, close to 1 second on most files).
+- Web Extension! working (using dart2js, formatter server is not used), functionality is almost the same, (~4x slower).
 - Improved performance of the formatter by ~15%.
 
 ### Fixed